fix: small quarts in sjb keymap tapdance code.
What I want is to be able to keep the current mods state event if one
of the keycode in the key pair of the tapdance has an a mod associated
with it.

example:

If we enter a tap-dance event with KC_LSFT mod active and we activate
the S(KC_0) for ")" in a tap-dance. Then the Shifted mod will be
removed when we are done the tapdance. We are using the macro_mods to
store the mod of the keycode if one exist and then remove it when we
are done. Didn't think it was worth creating a new mods state for this.